refactoring, react router added, dropdown certification url list working, contract updated to include id, hpa retrieval by id

+++ truffle/contracts/HPA.sol
@@ -51,7 +51,7 @@ contract HPA  {
        hpas_count[msg.sender]++;
     }
 
-    function createSimpleHPA(uint256 id, string memory firstName, string memory subject, string memory signature) public {
+    function createSimpleHPA(uint id, string memory firstName, string memory subject, string memory signature) public {
             uint simple_hpa_count = simple_hpas_count[msg.sender];
             simple_hpas[msg.sender][simple_hpa_count] = SimplePA(id, firstName, subject, signature);
             emit Certified(msg.sender);
